Make the SVG output ordered
The `HashMap`s for nucl, elem, and magic were changed to `Vec`s so that
the order was preserved. This means that the output is almost identical
when running with identical inputs. The color definitions are not
ordered, however. That would prevent accessing the colors with the color
name as a key. I don't think it's that important. It might be changed
later, though.